Initial Release
I just smacked together a zip file for the initial release of the HW-project files. I might start working away at chalking up some semblance of a 'menu-like' UI system for widgets, wizards, and whatnots but, this release stands alone as a proof of concept for the project, even so as it is 'keyboard-heavy' for the serious user. The best documentation is within the program by pressing 'F1', and gives a basic key description.